Create WinTeam CSV Clean
WinTeam exports invoice registers in CSV format, but duplicates all lines that fall under the same invoice register. A pivotal part of analysis is to identify recurring billing from one time billing. Normally this is a monthly task which takes over a day. This makes automated analysis difficult as a pivot table will report extremely inflated numbers as some invoices have large numbers of sub invoices, all being billed at the total invoice amount. This script will remove duplicates based on invoice number, and then identify which invoice number is recurring or one time based on a keyword search. This is done via formula seen on line 70. It will check for any keyword in the table and search for any match or partial match. Historically, billing on the first of the month is a sure indicator of a recurring charge, and is the word "revised". Then it moves the three most important columns to the far left so that a pivot table is done with the first three columns. The "Recurring or One Time" column should be used as the filter.